fix(tauri): Extract IPC types to shared module for Windows compatibility
The Windows build was failing because tcp_ipc was trying to import types
from the ipc module, which is Unix-only (#[cfg(unix)]). This caused:
  error[E0432]: unresolved import `crate::ipc`

Solution:
- Created ipc_common.rs with IpcCommand, IpcResponse, and process_command
- This module compiles on all platforms (no cfg restrictions)
- Both ipc.rs (Unix socket) and tcp_ipc.rs (TCP) now import from ipc_common
- Removed duplicate code from ipc.rs

This fixes the Windows release build failure in v0.5.1.
